requirements.txt
pystan
pandas
cython
matplotlib
arviz